requirements.txt
awscli==1.16.143
boto3==1.9.133
botocore==1.12.133
colorama==0.3.9
cycler==0.10.0
Cython==0.29.6
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
kiwisolver==1.0.1
matplotlib==3.0.3
mysqlclient==1.4.1
numpy==1.16.0
pandas==0.24.0
psycopg2==2.7.4
pyasn1==0.4.5
pyparsing==2.4.0
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
urllib3==1.24.2
